feat: FISE (Falken Interplanetary Settlement Engine) - Phase 1 Implementation
Smart Contracts:
- Add FiseEscrow.sol for JavaScript game logic settlement
- Modify MatchEscrow.sol: skip isValidMove for FISE matches, make _resolveRound virtual
- Add LogicRegistry updates for FISE support
- Remove deprecated SimpleDice logic and tests

Bots & Agents:
- Update HouseBot.ts: Add FISE detection, fix multiple match creation bug, use uint256 for hash
- Update SimpleAgent.ts: Add FISE support, fix hash calculation

Falken VM:
- Add complete Falken VM package for off-chain JS execution
- Implement Watcher, Referee, Fetcher, Reconstructor, Settler modules
- JS code transformation for sandbox execution (WIP - export{n as default} bug)

Indexer:
- Update for FISE event handling and Supabase sync

Dashboard:
- UI improvements and FISE match support

Docs:
- Add TECHNICAL_SPEC.md, TOKENOMICS.md, VISION.md
- Archive old docs

Scripts:
- Add debug, fix, and test scripts for FISE settlement

Handoff:
- Add CLAUDE_HANDOFF_FISE.md with current state and blockers

+++ scripts/commit-move.js
@@ -41,8 +41,9 @@ async function main() {
   const salt = ethers.hexlify(ethers.randomBytes(32));
   
   // Use FALKEN_V1 Domain Separator + address(this) for security
+  // IMPORTANT: Must use uint256 for round and move to match contract
   const hash = ethers.solidityPackedKeccak256(
-    ['string', 'address', 'uint256', 'uint8', 'address', 'uint8', 'bytes32'],
+    ['string', 'address', 'uint256', 'uint256', 'address', 'uint256', 'bytes32'],
     ["FALKEN_V1", escrowAddress, matchId, round, wallet.address, move, salt]
   );
